Item Tested on Page Module

add New Page
Update New Page
Delete new Page

Add New Page test
Field have to required

Update Page
was Adding new record instead of updating --Fixed
Redirecting to Logout Unscribction was not implemented On destroy not used

also pagelist was not showing arabic page name

Delete case checked as well Attaching Video for case reference

+++ Service/Concrete/PagesService.cs
@@ -39,19 +39,22 @@ namespace Service.Concrete
             try
             {
                 PagesModel pageModel = this.GetpagesById(Id);
-                Pages pages = new Pages()
+                if (pageModel != null)
                 {
-                    PageName = pageModel.PageName,
-                    PageNameAr = pageModel.PageNameAr,
-                    PageModuleEn=pageModel.PageModuleEn,
-                    PageModuleAr=pageModel.PageModuleAr,
-                    CreatedBy = pageModel.CreatedBy,
-                    CreatedDate = pageModel.CreatedDate,
-                    Id = pageModel.Id,
-                    Deleted = true
-                };
-                _pagesRepository.Delete(pages,userName);
-                _pagesRepository.Save();
+                    Pages pages = new Pages()
+                    {
+                        PageName = pageModel.PageName,
+                        PageNameAr = pageModel.PageNameAr,
+                        PageModuleEn = pageModel.PageModuleEn,
+                        PageModuleAr = pageModel.PageModuleAr,
+                        CreatedBy = pageModel.CreatedBy,
+                        CreatedDate = pageModel.CreatedDate,
+                        Id = pageModel.Id,
+                        Deleted = true
+                    };
+                    _pagesRepository.Delete(pages, userName);
+                    _pagesRepository.Save();
+                }
                 return true;
             }
             catch (Exception)
